Integrate cskau's keymapping command into main
