+++ src/epomakercontroller/keyboard_keys.py
@@ -3,13 +3,13 @@ from typing import Iterator
 from .configs.configs import Config
 
 
-@dataclass(frozen=True)
+@dataclass
 class KeyboardKey:
     name: str
     value: int
     display_str: str | None = None
 
-    def __post_init__(self):
+    def __post_init__(self) -> None:
         # If no display string, just use the name
         if not self.display_str:
             self.display_str = self.name
@@ -18,6 +18,7 @@ class KeyboardKey:
 class KeyboardKeys:
     """This class holds each keyboard key index along with it's name and display string"""
     def __init__(self, config: Config) -> None:
+        assert config.data is not None, "ERROR: Config has no data"
         self.all_keys = [KeyboardKey(**key) for key in config.data]
         self.name_to_key_dict = {}
         for key in self.all_keys:
